% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generalized_norm}
\alias{generalized_norm}
\title{Computes Quadratic Form Distance}
\usage{
generalized_norm(X, W, p = 1)
}
\arguments{
\item{X}{Numeric matrix of size \eqn{n\times K}{n*K}.}

\item{W}{Numeric matrix of size \eqn{K\times K}{K*K}.}

\item{p}{Numeric scalar.}
}
\description{
Computes Quadratic Form Distance
}
