% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{group_duplicates}
\alias{group_duplicates}
\alias{group_duplicates.data.frame}
\alias{group_duplicates.matrix}
\title{Identify duplicated rows and group them keeping track of memberships}
\usage{
group_duplicates(x, Treat = NULL, ids = NULL)

\method{group_duplicates}{data.frame}(x, Treat = NULL, ids = NULL)

\method{group_duplicates}{matrix}(x, Treat = NULL, ids = NULL)
}
\arguments{
\item{x}{Either a Numeric matrix or a data-frame.}

\item{Treat}{Integer vector of size \eqn{n}. Group indicator. Can be either -1 (no group), 0 (control) or 1 (treatment).}

\item{ids}{Integer vector with ids to be mapped to the set (internal use).}
}
\value{
A nested list with the following elements:
\itemize{
\item \code{groups} Lists of integers indicating the ids (row position) of
the group members.
\item \code{selected} Integer vector of length \code{length(groups)} with a proposed
set of ids to be used for the matching.
}

If \code{Treat} is passed, then it will return a list of length 3 with nested
lists as described above.
}
\description{
A lot of times observations may have duplicates within a same group.
For example, binary covariates have a high probability to be repeated.
Internally, \link{blop} controls for this and when performing the matching
takes into account duplicates so that the weights are equally distributed
accross duplicates.
}
\examples{
data(nsw)
group_duplicates(nsw[1:10, c("black", "hispanic", "nodegree", "re75")])
group_duplicates(nsw[1:10, c("black", "hispanic", "nodegree", "re75")], Treat = nsw$treat)
}
