% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matching_group}
\alias{matching_group}
\title{Finds matching group}
\usage{
matching_group(X, Treat, zeroindex = 1L)
}
\arguments{
\item{X}{Numeric matrix of size \eqn{n\times K}{n*K}.}

\item{Treat}{Integer vector of size \eqn{n}. Group indicator. Can be either -1 (no group), 0 (control) or 1 (treatment).}
}
\value{
A list of length \eqn{n} specifying for each i to which individuals
each j != i will be matched.
}
\description{
Finds matching group
}
