% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tag_duplicates}
\alias{tag_duplicates}
\title{Preprocess data identifying duplicates
This function identifies duplicates, groups them up and returns a vector
of indices of the observations to be used in the model, i.e. a new dataset
without duplicates.}
\usage{
tag_duplicates(x, zeroindex = TRUE)
}
\arguments{
\item{x}{Numeric matrix. Covariates over which duplicates should be tagged.}

\item{zeroindex}{Logical. Internal use.}
}
\value{
A list with the following elements:
- `groups` A list with the ids of duplicates grouped up.
- `selected_ids` An integer vector with the ids to be used.
- `n` `length(selected_ids)`.
}
\description{
Preprocess data identifying duplicates
This function identifies duplicates, groups them up and returns a vector
of indices of the observations to be used in the model, i.e. a new dataset
without duplicates.
}
