% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{teffect}
\alias{teffect}
\alias{teffect.default}
\alias{teffect.blopmatch_match}
\title{Treatment Effect Estimators using BLOP}
\usage{
teffect(X, y, effect = "ate", ...)

\method{teffect}{default}(X, y, Treat, effect = "ate", ...)

\method{teffect}{blopmatch_match}(x, y, effect = "ate", X = NULL, ...)
}
\arguments{
\item{X}{Numeric matrix of size \eqn{n\times K}{n*K}.}

\item{y}{Numeric vector of length \eqn{n}. Dependent variable.}

\item{Treat}{Integer vector of size \eqn{n}. Group indicator. Can be either -1 (no group), 0 (control) or 1 (treatment).}

\item{x}{An object of class \code{\link[blop]{blopmatch_match}}}
}
\description{
Treatment Effect Estimators using BLOP

\code{}
}
\author{
George G. Vega Yon
}
