% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blopmatch.r
\name{blop}
\alias{blop}
\alias{blopi_lpsolve}
\alias{blopi_glpk}
\title{Bilevel Optimization Problem Matching Solvers}
\usage{
blop(X, solver = "glpk")

blopi_lpsolve(xi, X)

blopi_glpk(xi, X)
}
\arguments{
\item{X}{A matrix of size \eqn{N\times K}{N * K}. Matching parameters.}

\item{solver}{A character scalar. Either \code{"glpk"} or \code{"lpsolve"}.}

\item{xi}{Numeric vector of length \eqn{k}. Ith individual covariates.}
}
\value{
In the case of \code{blop}, a list of class \code{blopmatch_match}:

\item{matches}{A list of size \code{N} with the BLOP solutions}
\item{relaxed}{An integer vector indicating in which of the \eqn{n} the
constraints were not binding (infeasibility of the problem).}
\item{X}{Original matrix \code{X} passed to the function.}
\item{X_pred}{A matrix of size \eqn{N\times K}{N * K} with the predicted values.}
}
\description{
\code{blop} matches all rows \eqn{i} in \code{X} to \eqn{j \neq i}{j!=i}.
This function is a wrapper of both \code{blopi_lpsolve} and \code{blopi_glpk}.

\code{blopi_lpsolve} solves the problem using the \pkg{lpSolverAPI} R package
which implements the lp_solver library.

\code{blopi_glpk} solves the problem using the \pkg{Rglpk} R package
which implements the GNU Linear Programming Kit.
}
\examples{
set.seed(123)

X <- matrix(rnorm(10*2), ncol=2)
ans <- blop(X)

plot(ans)

}
\author{
George G. Vega Yon
}
