% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_group.r
\name{matching_group}
\alias{matching_group}
\title{Enumerates potential matches}
\usage{
matching_group(Treat, exact = NULL)
}
\arguments{
\item{Treat}{Integer vector of size \eqn{n}. Group indicator. Can be either -1 (no group), 0 (control) or 1 (treatment).}

\item{exact}{Numeric matrix with \eqn{n} rows. When specified, exact matching is performed on those variables (see \code{\link{matching_group}}).}
}
\value{
A list of length \eqn{n} specifying for each i to which individuals
each j != i will be matched.
}
\description{
Given a variable \code{Treat}, \code{matching_group} enumerates the individuals
towards each observation can be matched against. If \code{exact} is specified,
then it performes exact match (see examples).
}
\details{
In the most common case \code{Treat} will be either of \code{0} or
\code{1}, in which case individuals will be matched against their opposite
group. If \code{Treat == -1}, then the group constraint will be void.
}
\examples{
# Asigning matching groups for lalonde --------------------------------------

data(lalonde, package = "MatchIt")
dat <- lalonde

# Match (no exact)
m_noexact <- matching_group(dat$treat)

# How many matches?
table(sapply(m_noexact, length))
table(dat$treat)

# What if we ask exact match on black ---------------------------------------

m_exact <- matching_group(dat$treat, dat$black)

table(sapply(m_exact, length))
with(dat, table(treat, black))
}
